refactored favorites into separate view model
- removed corresponding args from navhost
- main activity viewmodel no longer in use?
