+++ app/src/main/java/com/example/jokeaday/ui/screens/favorites/FavoritesPresentation.kt
@@ -6,7 +6,9 @@ import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.State
+import androidx.compose.runtime.livedata.observeAsState
 import androidx.compose.ui.Modifier
+import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavController
 import com.example.data.database.JokeEntity
 import com.example.jokeaday.ui.reusableComposables.CustomScaffold
@@ -16,9 +18,9 @@ import com.example.jokeaday.ui.reusableComposables.TextBox
 @Composable
 fun FavoritesPresentation(
     navController: NavController,
-    jokes: State<List<JokeEntity>?>
 ) {
-    val jokesList: List<JokeEntity> = jokes.value ?: emptyList()
+    val viewModel = hiltViewModel<FavoritesPresentationViewModel>()
+    val jokesList = viewModel.jokesDBLiveData.observeAsState()
 
     CustomScaffold(
         navController = navController,
@@ -28,8 +30,8 @@ fun FavoritesPresentation(
         Column(modifier = Modifier.padding(it)) {
             SpacerSmallest()
             LazyColumn {
-                if (jokesList.isNotEmpty()) {
-                    items(jokesList) { joke ->
+                jokesList.value?.let {jokes ->
+                    items(jokes) { joke ->
                         TextBox(
                             text = joke.setup,
                             onClick = { navController.navigate("JokePresentation/${joke.apiId}") }
